+++ src/mapping_localization/src/publisher/odometry_publisher.cpp
@@ -22,7 +22,7 @@ OdometryPublisher::OdometryPublisher(ros::NodeHandle& nh,
 }
 
 void OdometryPublisher::Publish(const Eigen::Matrix4f& transform_matrix, double time) {
-    ros::Time ros_time((float)time);
+    ros::Time ros_time(time);
     PublishData(transform_matrix, ros_time);
 }
